linux-gen: buffer: remove buffer_inlines header file
Moved odp_buffer_inlines.h content to odp_buffer_internal.h
and removed the file. Other xxx_inlines.h headers do not
contain internal functions but only code for API function
inlining.
