+++ platform/linux-generic/Makefile.am
@@ -86,7 +86,6 @@ noinst_HEADERS = \
 		  include/odp_atomic_internal.h \
 		  include/odp_bitmap_internal.h \
 		  include/odp_bitset.h \
-		  include/odp_buffer_inlines.h \
 		  include/odp_buffer_internal.h \
 		  include/odp_classification_datamodel.h \
 		  include/odp_classification_inlines.h \
