+++ platform/linux-generic/odp_ipsec_events.c
@@ -10,7 +10,6 @@
 #include <odp/api/shared_memory.h>
 
 #include <odp_buffer_internal.h>
-#include <odp_buffer_inlines.h>
 #include <odp_debug_internal.h>
 #include <odp_ipsec_internal.h>
 #include <odp_pool_internal.h>
