+++ platform/linux-generic/odp_queue_basic.c
@@ -14,7 +14,6 @@
 #include <odp/api/buffer.h>
 #include <odp_buffer_internal.h>
 #include <odp_pool_internal.h>
-#include <odp_buffer_inlines.h>
 #include <odp_internal.h>
 #include <odp/api/shared_memory.h>
 #include <odp/api/schedule.h>
