+++ platform/linux-generic/odp_schedule_scalable.c
@@ -25,7 +25,6 @@
 #include <odp_ishmpool_internal.h>
 
 #include <odp_align_internal.h>
-#include <odp_buffer_inlines.h>
 #include <odp_llqueue.h>
 #include <odp_queue_scalable_internal.h>
 #include <odp_schedule_if.h>
